requirements.txt
###############################################################################
# Runtime dependencies (production image)
# Pin (>= minor) for security & reproducibility; loosen if needed later.
###############################################################################
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
typing_extensions>=4.12.2
google-cloud-documentai==2.23.0
google-auth==2.40.0
google-api-core==2.25.1
googleapis-common-protos==1.70.0
# protobuf version pinned via constraints.txt (protobuf<5.0.0) to retain
# compatibility with current google-cloud-* libraries. Do NOT add an
# overlapping spec here to avoid conflicts; update constraints instead.
protobuf
tenacity==8.2.3
openai==1.93.3
httpx==0.27.2
prometheus-client==0.20.0
google-cloud-logging==3.10.0
google-api-python-client>=2.100.0
google-cloud-storage==2.18.0
PyPDF2==3.0.1

# NOTE: Removed unused libs (jinja2, backoff, cryptography, asgi-correlation-id, aiobreaker,
# pybreaker, gspread, slowapi, opentelemetry*, firestore, storage, doc/spreadsheet generators, PyMuPDF).
# Add back only when code introduces direct references.

